+++ common/src/main/java/fudge/notenoughcrashes/gui/ProblemScreen.java
@@ -5,6 +5,7 @@ import fudge.notenoughcrashes.gui.util.Widget;
 import fudge.notenoughcrashes.patches.PatchedCrashReport;
 import fudge.notenoughcrashes.platform.CommonModMetadata;
 import fudge.notenoughcrashes.utils.CrashLogUpload;
+import fudge.notenoughcrashes.utils.NecLocalization;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.MinecraftClient;
@@ -12,17 +13,17 @@ import net.minecraft.client.gui.screen.ConfirmChatLinkScreen;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.gui.widget.ButtonWidget;
 import net.minecraft.client.gui.widget.ClickableWidget;
-import net.minecraft.client.resource.language.I18n;
 import net.minecraft.client.util.math.MatrixStack;
-import net.minecraft.text.*;
+import net.minecraft.text.ClickEvent;
+import net.minecraft.text.LiteralText;
+import net.minecraft.text.MutableText;
+import net.minecraft.text.Text;
 import net.minecraft.util.Util;
 import net.minecraft.util.crash.CrashReport;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.io.File;
-import java.lang.reflect.Field;
-import java.net.URI;
 import java.util.*;
 
 @Environment(EnvType.CLIENT)
@@ -61,20 +62,22 @@ public abstract class ProblemScreen extends Screen {
         Set<CommonModMetadata> suspectedMods = ((PatchedCrashReport) report).getSuspectedMods();
 
         //TODO: this shouldn't be possible to be null anymore
-        if (suspectedMods == null) return new TranslatableText("notenoughcrashes.crashscreen.identificationErrored");
+        if (suspectedMods == null)
+            return NecLocalization.translatedText("notenoughcrashes.crashscreen.identificationErrored");
 
 
         // Minecraft exists and basically any stack trace, and loader exists in any launch,
         // it's better not to include them in the list of mods.
-        suspectedMods.removeIf(mod -> IGNORED_MODS.contains(mod.getId()));
+        suspectedMods.removeIf(mod -> IGNORED_MODS.contains(mod.id()));
 
-        if (suspectedMods.isEmpty()) return new TranslatableText("notenoughcrashes.crashscreen.noModsErrored");
+        if (suspectedMods.isEmpty())
+            return NecLocalization.translatedText("notenoughcrashes.crashscreen.noModsErrored");
 
         Text text = suspectedMods.stream()
-                .sorted(Comparator.comparing(CommonModMetadata::getName))
+                .sorted(Comparator.comparing(CommonModMetadata::name))
                 .map(mod -> {
-                    String issuesPage = mod.getIssuesPage();
-                    MutableText modText = new LiteralText(mod.getName());
+                    String issuesPage = mod.issuesPage();
+                    MutableText modText = new LiteralText(mod.name());
                     if (issuesPage != null) {
                         modText.styled(style -> style.withClickEvent(new ClickEvent(ClickEvent.Action.OPEN_URL, issuesPage)));
                     }
@@ -97,7 +100,7 @@ public abstract class ProblemScreen extends Screen {
     public void init() {
         widgets = new ArrayList<>();
 
-        this.getLinkButton = new ButtonWidget(width / 2 - 155 + 160, height / 4 + 120 + 12, 150, 20, new TranslatableText("notenoughcrashes.gui.getLink"),
+        this.getLinkButton = new ButtonWidget(width / 2 - 155 + 160, height / 4 + 120 + 12, 150, 20, NecLocalization.translatedText("notenoughcrashes.gui.getLink"),
                 buttonWidget -> {
                     try {
                         if (hasteLink == null) {
@@ -117,7 +120,7 @@ public abstract class ProblemScreen extends Screen {
                         }, hasteLink, true));
                     } catch (Throwable e) {
                         LOGGER.error("Exception when crash menu button clicked:", e);
-                        buttonWidget.setMessage(new TranslatableText("notenoughcrashes.gui.failed"));
+                        buttonWidget.setMessage(NecLocalization.translatedText("notenoughcrashes.gui.failed"));
                         buttonWidget.active = false;
                     }
                 });
@@ -149,7 +152,7 @@ public abstract class ProblemScreen extends Screen {
 
     protected void drawFileNameString(MatrixStack matrixStack, int y) {
         String fileNameString = report.getFile() != null ? "\u00A7n" + report.getFile().getName()
-                : I18n.translate("notenoughcrashes.crashscreen.reportSaveFailed");
+                : NecLocalization.localize("notenoughcrashes.crashscreen.reportSaveFailed");
         int stLen = textRenderer.getWidth(fileNameString);
         xLeft = width / 2 - stLen / 2;
         xRight = width / 2 + stLen / 2;