+++ common/src/main/java/fudge/notenoughcrashes/mixins/client/MixinMinecraftClient.java
@@ -4,13 +4,11 @@ import fudge.notenoughcrashes.NotEnoughCrashes;
 import fudge.notenoughcrashes.mixinhandlers.EntryPointCatcher;
 import fudge.notenoughcrashes.mixinhandlers.InGameCatcher;
 import fudge.notenoughcrashes.stacktrace.CrashUtils;
-import fudge.notenoughcrashes.utils.GlUtil;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.util.crash.CrashReport;
 import net.minecraft.util.thread.ReentrantThreadExecutor;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Overwrite;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -20,8 +18,8 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 import java.util.Queue;
 
+
 @Mixin(MinecraftClient.class)
-@SuppressWarnings("StaticVariableMayNotBeInitialized")
 public abstract class MixinMinecraftClient extends ReentrantThreadExecutor<Runnable> {
     @Shadow
     private CrashReport crashReport;
@@ -41,6 +39,8 @@ public abstract class MixinMinecraftClient extends ReentrantThreadExecutor<Runna
 
     @Inject(method = "run()V", at = @At(value = "FIELD", target = "Lnet/minecraft/client/MinecraftClient;crashReport:Lnet/minecraft/util/crash/CrashReport;"))
     private void onRunLoop(CallbackInfo ci) {
+        if (!NotEnoughCrashes.ENABLE_GAMELOOP_CATCHING) return;
+
         if (this.crashReport != null) {
             NotEnoughCrashes.logDebug("Handling run loop crash");
             InGameCatcher.handleServerCrash(crashReport);
@@ -54,40 +54,29 @@ public abstract class MixinMinecraftClient extends ReentrantThreadExecutor<Runna
     // Can't capture arg in inject so captured here
     @ModifyArg(method = "run()V", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/MinecraftClient;printCrashReport(Lnet/minecraft/util/crash/CrashReport;)V", ordinal = 1))
     private CrashReport atTheEndOfFirstCatchBeforePrintingCrashReport(CrashReport report) {
+        if (!NotEnoughCrashes.ENABLE_GAMELOOP_CATCHING) return report;
+
         NotEnoughCrashes.logDebug("Handling client game loop try/catch crash in first catch block");
         // we MUST use the report passed as parameter, because the field one only gets assigned in integrated server crashes.
-        InGameCatcher.handleClientCrash(report, renderTaskQueue);
+        InGameCatcher.handleClientCrash(report);
         return report;
     }
 
     // Can't capture arg in inject so captured here
     @ModifyArg(method = "run()V", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/MinecraftClient;printCrashReport(Lnet/minecraft/util/crash/CrashReport;)V", ordinal = 2))
     private CrashReport atTheEndOfSecondCatchBeforePrintingCrashReport(CrashReport report) {
+        if (!NotEnoughCrashes.ENABLE_GAMELOOP_CATCHING) return report;
+
         NotEnoughCrashes.logDebug("Handling client game loop try/catch crash in second catch block");
         // we MUST use the report passed as parameter, because the field one only gets assigned in integrated server crashes.
-        InGameCatcher.handleClientCrash(report, renderTaskQueue);
+        InGameCatcher.handleClientCrash(report);
         return report;
     }
-//
-//    @Inject(method = "run()V",
-//            at = @At(value = "INVOKE_ASSIGN",
-//                    target = "Lnet/minecraft/client/MinecraftClient;addDetailsToCrashReport(Lnet/minecraft/util/crash/CrashReport;)Lnet/minecraft/util/crash/CrashReport;"),
-//            cancellable = true)
-//    private void afterCrashHandled(CallbackInfo ci) {
-//        // Can't cancel in modifyarg so canceled here
-//        ci.cancel();
-//        // Continue game loop
-//        MinecraftClient.getInstance().run();
-//    }
-//
-//    @ModifyArg(method = "run()V",
-//            at = @At(value = "INVOKE",
-//                    target = "Lnet/minecraft/client/MinecraftClient;addDetailsToCrashReport(Lnet/minecraft/util/crash/CrashReport;)Lnet/minecraft/util/crash/CrashReport;"))
-//    private CrashReport onCrash(CrashReport report) {
-//        // Can't capture arg in inject so captured here
-//        InGameCatcher.handleClientCrash(report, renderTaskQueue);
-//        return report;
-//    }
+
+    @Inject(method = "cleanUpAfterCrash()V", at = @At("HEAD"))
+    private void beforeCleanUpAfterCrash(CallbackInfo info) {
+        InGameCatcher.cleanupBeforeMinecraft(renderTaskQueue);
+    }
 
     /**
      * Prevent the integrated server from exiting in the case it crashed
@@ -97,20 +86,4 @@ public abstract class MixinMinecraftClient extends ReentrantThreadExecutor<Runna
         CrashUtils.outputReport(report);
     }
 
-//
-//    /**
-//     * @author runemoro
-//     * @reason Disconnect from the current world and free memory, using a memory reserve
-//     * to make sure that an OutOfMemory doesn't happen while doing this.
-//     * <p>
-//     * Bugs Fixed:
-//     * - https://bugs.mojang.com/browse/MC-128953
-//     * - Memory reserve not recreated after out-of memory
-//     */
-//    @Inject(method = "cleanUpAfterCrash", at = @At("HEAD"))
-//    // can be replaced by 2-4 injection/redirections
-//    public void cleanUpAfterCrash(CallbackInfo ci) {
-//        GlUtil.resetState();
-////        InGameCatcher.resetGameState(renderTaskQueue);
-//    }
 }